Use line instead of stockrecord to determine price
Product.stockrecord was a deprecated helper to get the first stockrecord
for a product, so it was potentially the wrong stockrecord to look at
anyway. Using the price stored on the basket line is the correct thing
to do, and avoids the AttributeError as Product.stockrecord has since
been removed.